separate core and flow submodules. encapsulate
