+++ engine/runtime/expr/Cargo.toml
@@ -4,6 +4,8 @@ version = "0.0.1"
 edition = "2021"
 
 [dependencies]
+reearth-flow-types.workspace = true
+
 lalrpop-util = "0.22"
 logos = "0.15"
 serde_json.workspace = true
