+++ engine/runtime/expr/src/core/grammar.lalrpop
@@ -1,38 +1,38 @@
-use crate::ast::{BinOp, Expr, UnaryOp};
+use crate::core::ast::{BinOp, Expr, UnaryOp};
 
 grammar;
 
 extern {
     type Location = usize;
-    type Error = crate::error::Error;
+    type Error = crate::core::error::Error;
 
-    enum crate::lexer::Token {
-        "int"   => crate::lexer::Token::Int(<i64>),
-        "float" => crate::lexer::Token::Float(<f64>),
-        "str"   => crate::lexer::Token::Str(<String>),
-        "ident" => crate::lexer::Token::Ident(<String>),
-        "true"  => crate::lexer::Token::True,
-        "false" => crate::lexer::Token::False,
-        "null"  => crate::lexer::Token::Null,
-        "in"    => crate::lexer::Token::In,
-        "["     => crate::lexer::Token::LBracket,
-        "]"     => crate::lexer::Token::RBracket,
-        "("     => crate::lexer::Token::LParen,
-        ")"     => crate::lexer::Token::RParen,
-        ","     => crate::lexer::Token::Comma,
-        "+"     => crate::lexer::Token::Plus,
-        "-"     => crate::lexer::Token::Minus,
-        "*"     => crate::lexer::Token::Star,
-        "/"     => crate::lexer::Token::Slash,
-        "=="    => crate::lexer::Token::Eq,
-        "!="    => crate::lexer::Token::Ne,
-        "<"     => crate::lexer::Token::Lt,
-        "<="    => crate::lexer::Token::Le,
-        ">"     => crate::lexer::Token::Gt,
-        ">="    => crate::lexer::Token::Ge,
-        "&&"    => crate::lexer::Token::And,
-        "||"    => crate::lexer::Token::Or,
-        "!"     => crate::lexer::Token::Bang,
+    enum crate::core::lexer::Token {
+        "int"   => crate::core::lexer::Token::Int(<i64>),
+        "float" => crate::core::lexer::Token::Float(<f64>),
+        "str"   => crate::core::lexer::Token::Str(<String>),
+        "ident" => crate::core::lexer::Token::Ident(<String>),
+        "true"  => crate::core::lexer::Token::True,
+        "false" => crate::core::lexer::Token::False,
+        "null"  => crate::core::lexer::Token::Null,
+        "in"    => crate::core::lexer::Token::In,
+        "["     => crate::core::lexer::Token::LBracket,
+        "]"     => crate::core::lexer::Token::RBracket,
+        "("     => crate::core::lexer::Token::LParen,
+        ")"     => crate::core::lexer::Token::RParen,
+        ","     => crate::core::lexer::Token::Comma,
+        "+"     => crate::core::lexer::Token::Plus,
+        "-"     => crate::core::lexer::Token::Minus,
+        "*"     => crate::core::lexer::Token::Star,
+        "/"     => crate::core::lexer::Token::Slash,
+        "=="    => crate::core::lexer::Token::Eq,
+        "!="    => crate::core::lexer::Token::Ne,
+        "<"     => crate::core::lexer::Token::Lt,
+        "<="    => crate::core::lexer::Token::Le,
+        ">"     => crate::core::lexer::Token::Gt,
+        ">="    => crate::core::lexer::Token::Ge,
+        "&&"    => crate::core::lexer::Token::And,
+        "||"    => crate::core::lexer::Token::Or,
+        "!"     => crate::core::lexer::Token::Bang,
     }
 }
