+++ engine/runtime/expr/src/lib.rs
@@ -1,9 +1,19 @@
-pub mod ast;
-pub mod error;
-pub mod eval;
-pub(crate) mod lexer;
-pub mod parser;
+pub mod core;
+pub mod flow;
 
-pub use error::{Error, Result};
-pub use eval::{eval, Context};
-pub use parser::parse;
+pub use core::error::{Error, Result};
+pub use core::eval::Context;
+
+use serde_json::Value;
+
+/// Opaque handle to a compiled expression. Internals are not part of the public API.
+#[derive(Debug, Clone)]
+pub struct CompiledExpr(core::ast::Expr);
+
+pub fn compile(input: &str) -> Result<CompiledExpr> {
+    core::parser::parse(input).map(CompiledExpr)
+}
+
+pub fn eval(expr: &CompiledExpr, ctx: &Context) -> Result<Value> {
+    core::eval::eval(&expr.0, ctx)
+}
